In ProfileHeaderView, I added processing of a single click on the avatar.
When the click is triggered, the animation starts:
the avatar moves to the center of the screen and stretches across the width of the screen

a semi-ambiguous view is shown under the user's avatar, which overlaps the rest of the elements on the screen.

A button with a cross icon appears in the upper right corner.

Avatar and view animations run simultaneously and last 0.5 seconds.
The animation of the button will start after the avatar and view animation is completed and last 0.3 seconds;

When you click on the button with the cross icon, the reverse animation starts: